judge/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
mysql-connector-python==8.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
